v0.92
* improve typescript support
