+++ src/version.js
@@ -1,2 +1,2 @@
 // Generated by genversion.
-export const version = '0.90.0';
+export const version = '0.92.0';
