+++ src/web.js
@@ -1,6 +1,3 @@
 import litecanvas from './index.js'
 
-/**
- * @global
- */
 globalThis.litecanvas = litecanvas
